Revise README for clarity and consistency
Updated project title and refined project description for clarity. Adjusted wording to enhance readability and consistency throughout the README. Specifically, in removing the third-party organization name from the repo title, description, and content aside from the 'Acknowledgement' section, it reduces the risk associated with linking this repo to their work. This student project is meant as a demonstration now that it is a historical project done two years ago.